Delete README.md
This exercise utilized a few different JavaScript skills. One of the first and most important was learning and implementing the checkPageBounds function. I stored four different images within an arrary to show the moving pac man. I used a setInterval fucntion to allow the images to move after an if, else statement to determine direction and velocity.